requirements.txt
# Install with: pip(3) install -r requirements.txt --upgrade
GitPython
Mako<2.0.0
Sphinx
awesomeversion
bcrypt<5.0.0
cachecontrol<1
cachetools
cryptography
dacite
deepmerge
deprecated
dockerfile-parse>=1.2
docutils
ensure
flake8
gardener-oci
github3.py<5.0.0
html2text
jsonschema
jsonpath-ng
pycryptodome
pyjwt
pylama
pylint
pytest
python-dateutil
python-gitlab
python-keystoneclient
python-magic
python-swiftclient
pytimeparse
pyyaml
requests
ruamel.yaml==0.17.21
semver
slackclient<3.0.0
sphinx_rtd_theme
sseclient-py
tabulate
termcolor
urllib3
yamllint
